// src/app/types/CoinTypes.ts
export type Currencies = 'USD' | 'THB'

export type CoinContextType = {
  data: CoinState
  status: 'error' | 'success' | 'pending'
  currency: Currencies
  updateCurrency: (currency: Currencies) => void
}

export type CoinState = {
  trending: TrendingCoin[]
  coins: Coin[]
}

export type Coin = {
  id: string
  symbol: string
  name: string
  image: string
  current_price: number
  market_cap: number
  market_cap_rank: number
  fully_diluted_valuation: number
  total_volume: number
  high_24h: number
  low_24h: number
  price_change_24h: number
  price_change_percentage_24h: number
  market_cap_change_24h: number
  market_cap_change_percentage_24h: number
  circulating_supply: number
  total_supply: number
  max_supply: null | number
  ath: number
  ath_change_percentage: number
  ath_date: string
  atl: number
  atl_change_percentage: number
  atl_date: string
  roi: {
    times: number
    currency: string
    percentage: number
  } | null
  last_updated: string
  sparkline_in_7d: {
    price: number[]
  }
  price_change_percentage_1h_in_currency: number
}

export type TrendingCoin = {
  id: string
  coin_id: number
  name: string
  symbol: string
  market_cap_rank: number
  thumb: string
  small: string
  large: string
  slug: string
  price_btc: number
  score: number
}

export type SearchCoin = {
  id: string
  name: string
  api_symbol: string
  symbol: string
  market_cap_rank: number
  thumb: string
  large: string
}